fix: ExtractPayload가 제대로 작동하지 않는 에러 해결
