+++ src/main/java/umc/stockoneqback/global/annotation/ExtractPayload.java
@@ -1,5 +1,7 @@
 package umc.stockoneqback.global.annotation;
 
+import umc.stockoneqback.auth.domain.model.jwt.TokenType;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
