+++ src/main/java/umc/stockoneqback/global/config/WebConfig.java
@@ -5,7 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import umc.stockoneqback.auth.utils.JwtTokenProvider;
+import umc.stockoneqback.auth.utils.TokenProvider;
+import umc.stockoneqback.global.annotation.AuthArgumentResolver;
 import umc.stockoneqback.global.annotation.ExtractPayloadArgumentResolver;
 import umc.stockoneqback.global.annotation.ExtractTokenArgumentResolver;
 
@@ -14,7 +15,7 @@ import java.util.List;
 @Configuration
 @RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
-    private final JwtTokenProvider jwtTokenProvider;
+    private final TokenProvider tokenProvider;
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
@@ -28,7 +29,8 @@ public class WebConfig implements WebMvcConfigurer {
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new ExtractTokenArgumentResolver(jwtTokenProvider));
-        resolvers.add(new ExtractPayloadArgumentResolver(jwtTokenProvider));
+        resolvers.add(new ExtractTokenArgumentResolver(tokenProvider));
+        resolvers.add(new ExtractPayloadArgumentResolver(tokenProvider));
+        resolvers.add(new AuthArgumentResolver());
     }
 }
